requirements.txt
djangorestframework-jwt==1.11.0
djangorestframework==3.11.0
asgiref==3.2.7
Babel==2.8.0
Django==3.0.5
django-phonenumber-field==4.0.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.1.1
mysqlclient==1.4.6
phonenumbers==8.12.1
pytz==2019.3
django-localflavor==3.0.1
django-formtools==2.2
django-multiselectfield==0.1.12
django-cities-light==3.5.0

